add: crud pages for entities

+++ src/Controller/ProjectController.php
@@ -5,7 +5,6 @@ namespace App\Controller;
 use App\Entity\Project;
 use App\Form\ProjectType;
 use App\Repository\ProjectRepository;
-use App\Repository\ProjectsGroupRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -18,79 +17,65 @@ final class ProjectController extends AbstractController
     #[Route(name: 'app_project_index', methods: ['GET'])]
     public function index(ProjectRepository $projectRepository): Response
     {
-        $projects = $projectRepository->findAll();
-        return $this->json($projects, 200, [], ['groups' => ['project_read']]);
+        return $this->render('project/index.html.twig', [
+            'projects' => $projectRepository->findAll(),
+        ]);
     }
 
-    #[Route('/new', name: 'app_project_new', methods: ['POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager, ProjectsGroupRepository $projectsGroupRepository): Response
+    #[Route('/new', name: 'app_project_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
-        $data = json_decode($request->getContent(), true);
         $project = new Project();
-
-        $groupId = $data['projectGroup'] ?? null;
-
-        if ($groupId) {
-            $projectGroup = $projectsGroupRepository->find($groupId);
-
-            if ($projectGroup) {
-                $project->setProjectGroup($projectGroup);
-            } else {
-                return $this->json(['error' => 'Project group not found'], Response::HTTP_NOT_FOUND);
-            }
-        } else {
-            return $this->json(['error' => 'Project group ID is required'], Response::HTTP_BAD_REQUEST);
-        }
-
         $form = $this->createForm(ProjectType::class, $project);
-        $form->submit($data);
+        $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($project);
             $entityManager->flush();
 
-            return $this->json($project, Response::HTTP_CREATED, [], ['groups' => ['project_read']]);
+            return $this->redirectToRoute('app_project_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->json([
-            'data' => [
-                'errors' => (string)$form->getErrors(true, false),
-            ]
-        ], Response::HTTP_BAD_REQUEST);
+        return $this->render('project/new.html.twig', [
+            'project' => $project,
+            'form' => $form,
+        ]);
     }
 
     #[Route('/{id}', name: 'app_project_show', methods: ['GET'])]
     public function show(Project $project): Response
     {
-        return $this->json($project, 200, [], ['groups' => ['project_read']]);
+        return $this->render('project/show.html.twig', [
+            'project' => $project,
+        ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_project_edit', methods: ['PUT', 'PATCH'])]
+    #[Route('/{id}/edit', name: 'app_project_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Project $project, EntityManagerInterface $entityManager): Response
     {
-        $data = json_decode($request->getContent(), true);
-
         $form = $this->createForm(ProjectType::class, $project);
-        $form->submit($data, false);
+        $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
-            return $this->json($project, 200, [], ['groups' => ['project_read']]);
+
+            return $this->redirectToRoute('app_project_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->json([
-            'data' => [
-                'errors' => (string)$form->getErrors(true, false),
-            ]
-        ], Response::HTTP_BAD_REQUEST);
+        return $this->render('project/edit.html.twig', [
+            'project' => $project,
+            'form' => $form,
+        ]);
     }
 
-    #[Route('/{id}', name: 'app_project_delete', methods: ['DELETE'])]
-    public function delete(Project $project, EntityManagerInterface $entityManager): Response
+    #[Route('/{id}', name: 'app_project_delete', methods: ['POST'])]
+    public function delete(Request $request, Project $project, EntityManagerInterface $entityManager): Response
     {
-        $entityManager->remove($project);
-        $entityManager->flush();
+        if ($this->isCsrfTokenValid('delete'.$project->getId(), $request->getPayload()->getString('_token'))) {
+            $entityManager->remove($project);
+            $entityManager->flush();
+        }
 
-        return $this->json(['data' => null], Response::HTTP_NO_CONTENT);
+        return $this->redirectToRoute('app_project_index', [], Response::HTTP_SEE_OTHER);
     }
 }
